Disable hooks instead of accessing private Cite properties
I am not absolutely shure what I do here is really a working
alternative for what the code did before. I am suggesting this as a
much cleaner replacement for the previous hack, but ask others more
familiar with this extension to actually try it out. I have not tested
this suggestion.

Bug: T183323
Change-Id: Ib9e85b0eb590695ba2ea58ce30a34fe662fdb708