Add type hints to all class properties of DPLArticle
Somebody decided to initialize all fields with empty strings, even if
they must contain integers, or even Title objects. I decided to not
touch this for now, except for the title, which is exclusively set via
the constructor (and should be private, but again, I did not wanted to
touch this for now).

However, I did not added this string type to the documentation, mostly
because I could not find any code that actually uses these empty strings.

Change-Id: Ie2d2b448c2fd843d8018e3f3fa53f98f664c4a3c