Streamline a few overly complex regular expressions
This is done for performance as well as to improve the readability of
the code.

For example, replacing /^\s*/ can be super inefficient because this
pattern matches *all* lines, no matter if the line starts with a space
or not. It will replace nothing with nothing in all these cases. What
the pattern is really interested in are only lines that start with at
least one space. Therefor the +.

Change-Id: I13df1009ded1b45bc9e489e7ec0f383a4153e9dc