Fix whitespace-related code formatting issues
This patch does nothing but cleaning up code formatting issues, to make
this codebase conform more to the MediaWiki code style. It's often a
misplaced tab where a space should be, or a misplaced space before a tab.

In a few cases I try to split up long lines, but only if I had a reason
to touch the line anyway.

Change-Id: Ibff77a7af0c41eaa888d7cabf0b3c60fee2fe7db