Add PHP CodeSniffer checks to this codebase
I'm intentionally *not* starting to fix anything in this first patch,
but exclude all sniffs that have at least one violation in this
codebase. Fixes are for later patches.

Note I'm also intentionally excluding entire sniffs (A.B.C) instead
of individual errors (A.B.C.D), mostly to not have the list explode on
us. This should be changed one by one in later patches.

Change-Id: I508c37024786615be8fae54e3f6b6d64b9d9e8ea