Fix first round of PHP CodeSniffer violations
This fixes a whole bunch of PHPCS violations that are (mostly)
auto-fixable. I did a few very minor additional changes like moving a
comment inside of the { … } scope brackets it belongs to.

No actual changes are made to the code (e.g. no replacements of ==
with === or something like that).

I intentionally did *not* fixed indentions and alignments in this
patch, because there are so many violations. This is for a separate
patch.

Change-Id: Iec1c0a75f7155c15bbef9a76fc4e617ff2742da4