Fix warnings from static code analysis in DPL class
This fixes two relatively trivial warnings the static code analysis in
my PHPStorm reports. The empty strings (now null) are not used because
the parameter is an *out* parameter.

The uninitialized arrays are
usually not a problem because the loop that adds elements to the arrays
is usually not empty. However, all this makes the code more clean and
more reliable.

Change-Id: I6005b75ee439e51919521304edb5860f6140fae7