Mark public methods with the "public" keyword
I'm intentionally limiting this patch to methods that are actually
used from the outside, and must be public because of this.

I would love to make everything else private, but this is a bit more
error prone as it might break unknown code. I will continue working
on this in later patches.

Change-Id: I057414137481a7c3dcb3c123c71cce9ee5026799